auth/requirements.txt
sqlalchemy
asyncpg
fastapi
email-validator
python-jose
python-multipart